+++ packages/site/pages/api.js
@@ -1,19 +1,20 @@
 import { Fragment } from 'react';
-import Header from '~/templates/global/Header';
-import Footer from '~/templates/global/Footer';
-import content from '~/data/content.json';
-import menus from '~/data/menus.json';
 import GlobalTemplate from '~/templates/global/Template';
+import ContentPage from '~/templates/global-new/ContentPage';
+import { Section, PageHeader } from '~/templates/global-new/styled';
 import MenuPage from '~/templates/content/MenuPage';
 
 export default () => (
   <GlobalTemplate
-    title="API | Popmotion"
-    description="Popmotion is a functional, reactive JavaScript motion library."
+    title="API | Popmotion Pure"
+    description="Popmotion Pure is a functional, flexible JavaScript motion library."
     theme="pure"
   >
-    <Header section="api" />
-    <MenuPage section="api" title="API" />
-    <Footer />
+    <ContentPage section="api">
+      <Section>
+        <PageHeader>API</PageHeader>
+        <MenuPage section="api" />
+      </Section>
+    </ContentPage>
   </GlobalTemplate>
 );