+++ packages/site/pages/stylefire/api.js
@@ -1,9 +1,7 @@
 import { Fragment } from 'react';
-import Header from '~/templates/global/Header';
-import Footer from '~/templates/global/Footer';
-import content from '~/data/content.json';
-import menus from '~/data/menus.json';
 import GlobalTemplate from '~/templates/global/Template';
+import ContentPage from '~/templates/global-new/ContentPage';
+import { Section, PageHeader } from '~/templates/global-new/styled';
 import MenuPage from '~/templates/content/MenuPage';
 
 export default () => (
@@ -12,8 +10,11 @@ export default () => (
     description="Simple HTML & SVG styler, optimised for animation"
     theme="stylefire"
   >
-    <Header section="api" />
-    <MenuPage section="api" title="API" />
-    <Footer />
+    <ContentPage section="api">
+      <Section>
+        <PageHeader>API</PageHeader>
+        <MenuPage section="api" />
+      </Section>
+    </ContentPage>
   </GlobalTemplate>
 );